:art: Add tabs to problems page (#463)
